Initial check in of GPOne Projects
This check in contains 243 files. They span a number of different
game-related projects.

Included are the following:
- AnimationEditor: Sprite-based animation editor (In Progress)
- ItemEditor: Item library manager (In Progress)
- GameObjects: A C# library containing a number of GameObject classes,
also includes some WPF controls (In Progress)
- XnaControls: A C#/XNA library containing a number of classes which
closely mimic primitive controls in the WPF framework (In Progress)
- XnaGameObjects: A C#/XNA library containing a number of GameObject
classes (In Progress)
- GamepadInput: A sample WPF game which relies entirely on the other
included libraries. (As a proof of concept, it is complete but very
ugly)
- WPFXSample: A C#/WPF based tile editor. (In Progress)
-GPOne: A sandbox for testing classes in the C# libraries
- WindowsGameOne: XNA application for testing C# game libraries,
currently it demonstrates a C#/XNA UI Framework (In-Progress)
- WindowsGameOneContent: The companion project used by all XNA
applications to store and load content.